requirements.txt
requests
python-dotenv
pymupdf>=1.23.0